+++ src/main/java/de/chrgroth/jsonstore/AbstractJsonStore.java
@@ -1,12 +1,10 @@
-package de.chrgroth.jsonstore.store;
+package de.chrgroth.jsonstore;
 
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
-import de.chrgroth.jsonstore.json.JsonService;
-import de.chrgroth.jsonstore.storage.StorageService;
-import de.chrgroth.jsonstore.store.exception.JsonStoreException;
+import de.chrgroth.jsonstore.metrics.JsonStoreMetrics;
 
 /**
  * Represents a JSON store for a concrete class. Access is provided using delegate methods to Java built in stream API. You may use flexjson annotations to
@@ -20,30 +18,55 @@ import de.chrgroth.jsonstore.store.exception.JsonStoreException;
  */
 public abstract class AbstractJsonStore<T, P> {
 
-    protected JsonService jsonService;
-    protected StorageService storageService;
+    protected final JsonService jsonService;
+    protected final StorageService storageService;
 
-    protected JsonStoreMetadata<T, P> metadata;
+    protected final JsonStoreMetadata<T, P> metadata;
     protected final boolean autoSave;
     protected final Map<Integer, VersionMigrationHandler> migrationHandlers;
 
     protected AbstractJsonStore(JsonService jsonService, StorageService storageService, String uid, Class<T> payloadClass, Integer payloadTypeVersion, boolean singleton,
             boolean autoSave, VersionMigrationHandler... migrationHandlers) {
+
+        // set json service
+        if (jsonService == null) {
+            throw new JsonStoreException("jsonService must be provided!!");
+        }
         this.jsonService = jsonService;
+
+        // set storage service
+        if (storageService == null) {
+            throw new JsonStoreException("storageService must be provided!!");
+        }
         this.storageService = storageService;
+
+        // create empty metadata
         metadata = new JsonStoreMetadata<>();
         metadata.setUid(uid);
         metadata.setPayloadType(payloadClass.getName());
         metadata.setPayloadTypeVersion(payloadTypeVersion);
         metadata.setSingleton(singleton);
         metadata.setCreated(new Date());
+
+        // set auto-save mode
         this.autoSave = autoSave;
+
+        // set migration handlers
         this.migrationHandlers = new HashMap<>();
         if (migrationHandlers != null) {
             for (VersionMigrationHandler migrationHandler : migrationHandlers) {
                 this.migrationHandlers.put(migrationHandler.sourceVersion(), migrationHandler);
             }
         }
+    }
+
+    /**
+     * Returns the stores uid.
+     *
+     * @return uid
+     */
+    public String getUid() {
+        return metadata.getUid();
     }
 
     /**
@@ -59,36 +82,13 @@ public abstract class AbstractJsonStore<T, P> {
      * @return metrics, never null
      */
     public JsonStoreMetrics computeMetrics() {
-        return new JsonStoreMetrics(metadata.getUid(), metadata.getPayloadType(), size(), metadata.getModified(), storageService.storageSize(metadata));
-    }
-
-    /**
-     * Returns the stores uid.
-     *
-     * @return uid
-     */
-    public String getUid() {
-        return metadata.getUid();
-    }
-
-    /**
-     * Checks if store is persistent.
-     *
-     * @return true if store is persistent, false otherwise
-     */
-    public final boolean isPersistent() {
-        return storageService != null;
+        return new JsonStoreMetrics(metadata.getUid(), metadata.getPayloadType(), size(), metadata.getModified(), storageService.size(metadata));
     }
 
     /**
      * Saves all data contained in store to configured file. No action if store is not persistent.
      */
     public final void save() {
-
-        // abort on transient stores
-        if (!isPersistent()) {
-            return;
-        }
 
         // create JSON
         String json = toJson();
@@ -115,11 +115,6 @@ public abstract class AbstractJsonStore<T, P> {
      */
     public final void load() {
 
-        // abort on transient stores
-        if (!isPersistent()) {
-            return;
-        }
-
         // load
         String json = storageService.read(metadata);
 
@@ -128,7 +123,7 @@ public abstract class AbstractJsonStore<T, P> {
     }
 
     /**
-     * Creates store elements from given JSON data and replaces all store contents.Will invoke {@link #save()} if using auto-save mode. If any error occurs a
+     * Creates store elements from given JSON data and replaces all store contents. Will invoke {@link #save()} if using auto-save mode. If any error occurs a
      * {@link JsonStoreException} will be thrown cause otherwise data loss may occur on next successful save.
      *
      * @param json
@@ -138,18 +133,13 @@ public abstract class AbstractJsonStore<T, P> {
         fromJsonInternal(json, true);
     }
 
-    private void fromJsonInternal(String json, boolean forceStore) {
-
-        boolean migrated = jsonService.fromJson(metadata, migrationHandlers, json);
-
-        // TODO move to JsonService using come callback handler
-        metadataRefreshed();
-
-        // TODO move to JsonService using come callback handler
-        // save
-        if (autoSave && (forceStore || migrated)) {
-            save();
-        }
+    protected void fromJsonInternal(String json, boolean forceStore) {
+        jsonService.fromJson(metadata, migrationHandlers, json, migrated -> {
+            metadataRefreshed();
+            if (autoSave && (forceStore || migrated)) {
+                save();
+            }
+        });
     }
 
     /**
@@ -161,8 +151,6 @@ public abstract class AbstractJsonStore<T, P> {
      * Drops store file explicitly. Transient data in store remains unchanged.
      */
     public final void drop() {
-        if (isPersistent()) {
-            storageService.delete(metadata);
-        }
+        storageService.delete(metadata);
     }
 }