+++ src/main/java/de/chrgroth/jsonstore/json/flexjson/custom/DateTypeHandler.java
@@ -4,7 +4,6 @@ import java.lang.reflect.Type;
 import java.time.format.DateTimeFormatter;
 import java.util.Date;
 
-import de.chrgroth.jsonstore.json.flexjson.AbstractFlexjsonTypeHandler;
 import flexjson.ObjectBinder;
 
 /**